Got the Pipe detection and Line split working
Will detect and count all Pipe's in the line.  Will then separate out each portion before the pipe into an array.
The last command will be at array[0], second to last at array[1], so on until the first command is at array[n].  Thus we can run the pipe commands from array[0] to array[n] and have them be in the correct order of execution.