Updated Pipe Count
Removed unneeded code after discussing what I was doing with gondree.  He stated we didn't need to parse the line ourselves and his token was sufficient.  Removed the code for our parsing and updated pipeCount to work with his line_words array and his num_words to detect the number of Pipes.